refactor notification logic into custom hooks
